Move kernel_filesystems functionality into a little library.
Related: #12

This way, tasks can know what the kernel filesystems are if they need to.
Avoids circular imports of filesystems, task, then filesystem again.
